write test trying to sync changes between an entrys components

+++ backend/FwLite/FwLiteProjectSync.Tests/Fixtures/SyncFixture.cs
@@ -18,6 +18,8 @@ public class SyncFixture : IAsyncLifetime
     public CrdtFwdataProjectSyncService SyncService =>
         _services.ServiceProvider.GetRequiredService<CrdtFwdataProjectSyncService>();
     private readonly string _projectName;
+    private bool _crdtEnabled = true;
+    private bool _fwDataEnabled = true;
 
     public static SyncFixture Create(string projectName) => new(projectName);
 
@@ -41,23 +43,30 @@ public class SyncFixture : IAsyncLifetime
 
     public async Task InitializeAsync()
     {
-        var projectsFolder = _services.ServiceProvider.GetRequiredService<IOptions<FwDataBridgeConfig>>().Value
-            .ProjectsFolder;
-        if (Path.Exists(projectsFolder)) Directory.Delete(projectsFolder, true);
-        Directory.CreateDirectory(projectsFolder);
-        var lcmCache = _services.ServiceProvider.GetRequiredService<IProjectLoader>()
-            .NewProject($"{_projectName}.fwdata", "en", "fr");
-        var projectGuid = lcmCache.LanguageProject.Guid;
-        FwDataApi = _services.ServiceProvider.GetRequiredService<FwDataFactory>().GetFwDataMiniLcmApi(_projectName, false);
+        Guid projectGuid = Guid.NewGuid();
+        if (_fwDataEnabled)
+        {
+            var projectsFolder = _services.ServiceProvider.GetRequiredService<IOptions<FwDataBridgeConfig>>().Value
+                .ProjectsFolder;
+            if (Path.Exists(projectsFolder)) Directory.Delete(projectsFolder, true);
+            Directory.CreateDirectory(projectsFolder);
+            var lcmCache = _services.ServiceProvider.GetRequiredService<IProjectLoader>()
+                .NewProject($"{_projectName}.fwdata", "en", "fr");
+            projectGuid = lcmCache.LanguageProject.Guid;
+            FwDataApi = _services.ServiceProvider.GetRequiredService<FwDataFactory>()
+                .GetFwDataMiniLcmApi(_projectName, false);
+        }
 
-        var crdtProjectsFolder =
-            _services.ServiceProvider.GetRequiredService<IOptions<LcmCrdtConfig>>().Value.ProjectPath;
-        if (Path.Exists(crdtProjectsFolder)) Directory.Delete(crdtProjectsFolder, true);
-        Directory.CreateDirectory(crdtProjectsFolder);
-        var crdtProject = await _services.ServiceProvider.GetRequiredService<ProjectsService>()
-            .CreateProject(new(_projectName, projectGuid));
-        _services.ServiceProvider.GetRequiredService<ProjectContext>().Project = crdtProject;
-        CrdtApi = _services.ServiceProvider.GetRequiredService<IMiniLcmApi>();
+        if (_crdtEnabled)
+        {
+            var crdtProjectsFolder = _services.ServiceProvider.GetRequiredService<IOptions<LcmCrdtConfig>>().Value.ProjectPath;
+            if (Path.Exists(crdtProjectsFolder)) Directory.Delete(crdtProjectsFolder, true);
+            Directory.CreateDirectory(crdtProjectsFolder);
+            var crdtProject = await _services.ServiceProvider.GetRequiredService<ProjectsService>()
+                .CreateProject(new(_projectName, projectGuid));
+            _services.ServiceProvider.GetRequiredService<ProjectContext>().Project = crdtProject;
+            CrdtApi = _services.ServiceProvider.GetRequiredService<IMiniLcmApi>();
+        }
     }
 
     public async Task DisposeAsync()
@@ -67,6 +76,16 @@ public class SyncFixture : IAsyncLifetime
 
     public IMiniLcmApi CrdtApi { get; set; } = null!;
     public FwDataMiniLcmApi FwDataApi { get; set; } = null!;
+
+    public void DisableCrdt()
+    {
+        _crdtEnabled = false;
+    }
+
+    public void DisableFwData()
+    {
+        _fwDataEnabled = false;
+    }
 }
 
 public class MockProjectContext(CrdtProject project) : ProjectContext
